dsp/requirements.txt
pyserial
PyQt5
numpy
pandas
scipy
sklearn
mne
mne_qt_browser
pynput